--- a/config/config.exs
+++ b/config/config.exs
@@ -6,7 +6,6 @@ if config_env() == :test do
     tags: %{},
     enable_source_code_context: true,
     root_source_code_paths: [File.cwd!()],
-    source_code_exclude_patterns: [],
     hackney_opts: [recv_timeout: 50, pool: :sentry_pool],
     send_result: :sync,
     send_max_attempts: 1,