--- a/test/context_test.exs
+++ b/test/context_test.exs
@@ -77,7 +77,6 @@ defmodule Sentry.ContextTest do
     assert event.user == %{}
   end
 
-  @tag start_app: false
   test "passing in tags context as option overrides Context and Application config" do
     Context.set_tags_context(%{"key" => "345", "key1" => "123"})
     put_test_config(tags: %{"key" => "overridden", "key2" => "1234", "key3" => "12345"})