--- a/test/test_helper.exs
+++ b/test/test_helper.exs
@@ -5,3 +5,5 @@ File.rm_rf!(Sentry.Sources.path_of_packaged_source_code())
 ExUnit.after_suite(fn _ ->
   File.rm_rf!(Sentry.Sources.path_of_packaged_source_code())
 end)
+
+{:ok, _} = Plug.Cowboy.http(Sentry.ExamplePlugApplication, [], port: 8003)