Before:

```
Finished in 4.4 seconds (0.6s async, 3.8s sync)
12 doctests, 159 tests, 0 failures
```

After:

```
Finished in 0.9 seconds (0.4s async, 0.4s sync)
12 doctests, 159 tests, 0 failures
```